+++ couplet_composer/support/environment.py
@@ -74,7 +74,7 @@ def get_composing_directory(
         "{}-{}-{}-{}".format(
             target.system,
             target.machine,
-            build_variant,
+            build_variant.lower(),
             cmake_generator.replace(" ", "_")
         )
     )
@@ -111,7 +111,7 @@ def get_destination_directory(
         "{}-{}-{}-{}".format(
             target.system,
             target.machine,
-            build_variant,
+            build_variant.lower(),
             cmake_generator.replace(" ", "_")
         )
     )
@@ -179,7 +179,7 @@ def get_dependencies_directory(build_root, target, build_variant):
         build_root,
         "local",
         "lib",
-        "{}-{}-{}".format(target.system, target.machine, build_variant)
+        "{}-{}-{}".format(target.system, target.machine, build_variant.lower())
     )
 
 
@@ -203,7 +203,7 @@ def get_dependency_version_data_file(build_root, target, build_variant):
         "versions-{}-{}-{}".format(
             target.system,
             target.machine,
-            build_variant
+            build_variant.lower()
         )
     )